% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createModuleConfig}
\alias{createModuleConfig}
\title{createModuleConfig}
\usage{
createModuleConfig(
  connectionDetails,
  moduleId = "about",
  tabName = "About",
  shinyModulePackage = "OhdsiShinyModules",
  moduleUiFunction = "aboutViewer",
  moduleServerFunction = "aboutServer",
  moduleDatabaseConnectionKeyService = NULL,
  moduleDatabaseConnectionKeyUsername = NULL,
  moduleInfoBoxFile = "aboutHelperFile()",
  moduleIcon = "info",
  resultDatabaseDetails
)
}
\arguments{
\item{configList}{An R list with the config settings}

\item{configLocation}{The location to save the config json file}
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the config specification
}
\details{
User specifies the settings to create a config for a module
}
