% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShinyApp.R
\name{createShinyApp}
\alias{createShinyApp}
\title{createShinyApp}
\usage{
createShinyApp(config, shinyAppLocation, overwrite = F)
}
\arguments{
\item{config}{An R list with the specifications}

\item{shinyAppLocation}{The folder to create the shiny app in}

\item{overwrite}{Whether to overwrite the ui and sever files if they exist}
}
\value{
A shiny ui and server file in the specified shinyAppLocation location
}
\description{
Creates the ui and server files for the shiny app
}
\details{
Given a specification list, this fuction creates a shiny app for viewing OHDSI shiny analyses
}
