% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{viewShiny}
\alias{viewShiny}
\title{viewShiny}
\usage{
viewShiny(shinyAppLocation)
}
\arguments{
\item{shinyAppLocation}{The folder to create the shiny app in}
}
\value{
The shiny app will open
}
\description{
Open the shiny app
}
\details{
User specifies the shiny app location and this function opens it
}
