% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewShiny.R
\name{viewShiny}
\alias{viewShiny}
\title{viewShiny}
\usage{
viewShiny(config)
}
\arguments{
\item{config}{The json with the app config}
}
\value{
The shiny app will open
}
\description{
Open the shiny app
}
\details{
User specifies the shiny app location and this function opens it
}
